feat: 수입, 지출 필터체크박스 공통 컴포넌트로 관리
- history/component/HistoryScreen → component/Button.kt
